+++ Classes/Extensions.Identity.App.cs
@@ -9,6 +9,7 @@
 
 using Microsoft.Identity.Client;
 using System;
+using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 
 namespace Extensions.Identity
@@ -92,6 +93,98 @@ namespace Extensions.Identity
                 .WithCertificate(cert)
                 .WithAuthority(new Uri(authority))
                 .Build();
+        }
+
+        /// <summary>
+        /// Get a valid IPublicClientApplication.
+        /// </summary>
+        /// <param name="appId">The GUID for the App/Client ID.</param>
+        /// <param name="tenantString">The tenant string e.g. for 
+        /// 'contoso.sharepoint.com' the value would be 'contoso'.</param>
+        /// <param name="environmentName">The environment name enum value.
+        /// Defaults to O365USGovGCCHigh.</param>
+        /// <returns>An IPublicClientApplication object.</returns>
+        public static IPublicClientApplication GetApp(
+            string appId,
+            string tenantString,
+            EnvironmentName environmentName = EnvironmentName.O365USGovGCCHigh)
+        {
+            //Get the authority URL for the given environment.
+            string authority = GetAuthority(environmentName, tenantString);
+            //Create the app using appId, cert and authority.
+            var app = PublicClientApplicationBuilder.Create(appId)
+                .WithAuthority(new Uri(authority))
+                .WithDefaultRedirectUri()
+                .Build();
+            var accounts = app.GetAccountsAsync().GetAwaiter().GetResult();
+            if (GetPublicAppAuthResult(
+                ref app,
+                ref accounts,
+                PublicAppAuthResultType.Silent) == null)
+            {
+                if (GetPublicAppAuthResult(
+                    ref app,
+                    ref accounts,
+                    PublicAppAuthResultType.Interactive) == null)
+                {
+                    if (GetPublicAppAuthResult(
+                        ref app,
+                        ref accounts,
+                        PublicAppAuthResultType.Prompt) == null)
+                    {
+                        return null;
+                    }
+                }
+            }
+            return app;
+        }
+
+        /// <summary>
+        /// The different types of authentication efforts to use.
+        /// </summary>
+        internal enum PublicAppAuthResultType
+        {
+            Silent,
+            Interactive,
+            Prompt
+        }
+
+        /// <summary>
+        /// Get an IPublicClientApplication AuthenticationResult based on
+        /// different authentication methods from least to most intrusive.
+        /// </summary>
+        /// <param name="app">A reference to the IPublicClientApplication
+        /// that is being authenticated.</param>
+        /// <param name="accounts">A reference to the accounts associated
+        /// with the given IPublicClientApplication.</param>
+        /// <param name="authType">The type of authentication to use.  Values
+        /// range from least to most intrusive to the end user.</param>
+        /// <returns>A valid AuthenticationResult or null.</returns>
+        internal static AuthenticationResult? GetPublicAppAuthResult(
+            ref IPublicClientApplication app,
+            ref System.Collections.Generic.IEnumerable<IAccount> accounts,
+            PublicAppAuthResultType authType)
+        {
+            switch (authType)
+            {
+                case PublicAppAuthResultType.Silent:
+                    return app.AcquireTokenSilent(
+                        Scopes.Graph,
+                        accounts.FirstOrDefault())
+                        .ExecuteAsync().GetAwaiter().GetResult();
+                    break;
+                case PublicAppAuthResultType.Interactive:
+                    return app.AcquireTokenInteractive(Scopes.Graph)
+                        .ExecuteAsync().GetAwaiter().GetResult();
+                    break;
+                case PublicAppAuthResultType.Prompt:
+                    return app.AcquireTokenInteractive(Scopes.Graph)
+                        .WithAccount(accounts.FirstOrDefault())
+                        .WithPrompt(Prompt.SelectAccount)
+                        .ExecuteAsync().GetAwaiter().GetResult();
+                    break;
+            }
+            return null;
         }
 
         /// <summary>